Add setting to avoid running pip install
It's not typically good manners to be installing Python packages for the
user behind the scenes. Add an option the avoids the `pip install' when
set.

This variable might actually be better off defaulting to `t' since an
unexpected `pip install' with the user's system Python interpreter could
end up polluting their Python site package space. Alternatively, the
automatic `pip install' could be removed altogether since the Python
dependencies are already documented in the README.